split up docker client into multiple submodules
