+++ tests/unit/test_misc.py
@@ -12,7 +12,7 @@ from localstack.services.generic_proxy import ProxyListener, start_proxy_server
 from localstack.utils import async_utils, config_listener
 from localstack.utils.aws import aws_stack
 from localstack.utils.common import TMP_FILES, download, json_safe, load_file, now_utc, parallelize
-from localstack.utils.docker_utils import PortMappings
+from localstack.utils.container_utils.container_client import PortMappings
 from localstack.utils.http_utils import create_chunked_data, parse_chunked_data
